Perform cleanup of discussions for Azure Devops Decorator
The Azure Devops Decorator does not currently close comment threads where the issue no longer exists. As the concept of submitting comments to threads and reporting an overall pipeline status is similar between Gitlab and Azure Devops, the common co-ordination of retrieving existing threads, checking if they contain SonarQube comments, closing any outdated issues, reporting new issues, and submitting a pipeline result has been moved into a `DiscussionAwarePullRequestDecorator` with both the Gitlab and Azure DevOps decorator extending this base decorator and providing the implementations for the actual interactions and response handling for the relevant ALMs.

As there is no reliable way of obtaining the details of the current user details on the version of the Azure DevOps API used by the plugin, the filtering for messages posted by the SonarQube user presumes that any thread started with a message containing the `View in SonarQube` link is a SonarQube comment, and therefore includes that thread in any subsequent cleanup activities.

Due to the re-use of the code from the Gitlab decorator, the Azure DevOps decorator is now Monorepo aware, with pipeline results being submitted under individual project keys, and comment cleanup checking the project details contained in each comment.